corrigida delecao ao clicar em cancelar

+++ todo.frontend/todo/src/app/components/finalizados/finalizados.component.ts
@@ -37,15 +37,17 @@ export class FinalizadosComponent implements OnInit {
     this.dialogService
       .openConfirmDialog("Tem certeza que deseja deletar?")
       .afterClosed()
-      .subscribe((resposta) => {
-        this.service.delete(id).subscribe((resposta) => {
-          if (resposta === null) {
-            this.service.message("Tarefa deletada com sucesso");
-            this.listFinished = this.listFinished.filter(
-              (todo) => todo.id !== id
-            );
-          } else this.service.message("Algo deu errado");
-        });
+      .subscribe((resp) => {
+        if (resp === true) {
+          this.service.delete(id).subscribe((resposta) => {
+            if (resposta === null) {
+              this.service.message("Tarefa deletada com sucesso");
+              this.listFinished = this.listFinished.filter(
+                (todo) => todo.id !== id
+              );
+            } else this.service.message("Algo deu errado");
+          });
+        }
       });
   }
